Combine provider resource schema and traits Add a new trait for state properties not present in resource schema
